move buildAgentLoopArgs
